+++ authpass/lib/utils/path_utils.dart
@@ -6,7 +6,7 @@ import 'package:authpass/env/_base.dart';
 import 'package:authpass/utils/platform.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
-import 'package:path_provider/path_provider.dart';
+import 'package:path_provider/path_provider.dart' as path_provider;
 
 class PathUtils {
   factory PathUtils() => _instance;
@@ -17,15 +17,20 @@ class PathUtils {
   static final Completer<bool> runAppFinished = Completer<bool>();
   static Future<bool> get waitForRunAppFinished => runAppFinished.future;
 
+  Future<Directory> getTemporaryDirectory() async {
+    return _namespaced(await path_provider.getTemporaryDirectory());
+  }
+
   Future<Directory> getAppDataDirectory() async {
     if (AuthPassPlatform.isWeb) {
       throw UnsupportedError('Not supported on web.');
     }
     if (AuthPassPlatform.isIOS || AuthPassPlatform.isMacOS) {
-      return _namespaced(await getApplicationSupportDirectory());
+      return _namespaced(await path_provider.getApplicationSupportDirectory());
     }
     if (AuthPassPlatform.isAndroid) {
-      return _namespaced(await getApplicationDocumentsDirectory());
+      return _namespaced(
+          await path_provider.getApplicationDocumentsDirectory());
     }
     return _namespaced(await _getDesktopDirectory());
   }
@@ -37,6 +42,8 @@ class PathUtils {
     }
     return Directory(path.join(base.path, namespace));
   }
+
+  String get namespace => Env.value?.storageNamespace;
 
   Future<Directory> getLogDirectory() async {
     return Directory(